[spc.main] Log errors nicely w/ step & lamps first
Now, we say *how* the timepoint failed in the error log. Also, lights
happen before chambers, so that the plants don't get eclipse conditions
if the cabinets are down. This may/may not be a good idea, but they
chambers seem to intermittently fail more than they die, so the original
fear of cooking plants is probably invalid. And the lights will be
powered off if the chambers are fully off/dead anyway.